physical-ai-book/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
google-generativeai==0.8.4
qdrant-client==1.12.1
cohere==5.11.0
python-dotenv==1.0.1